Our first (failing) compression test

+++ lib/src/main/java/fsst/Symbol.java
@@ -31,8 +31,8 @@ public class Symbol {
 
     }
 
-    Symbol(int c, int code) {
-        // NOTE: This needs to be checked
+    Symbol(byte c, int code) {
+        // TODO: This needs to be checked
         this.icl = (1 << 28) | (code << 16) | 56;
         this.value = c;
     }
